Fixes #22 and tune multiplication thresholds

+++ examples/example002d_pi_limb08.cpp
@@ -55,12 +55,12 @@ bool math::wide_decimal::example002d_pi_limb08()
   using local_allocator_type = util::n_slot_array_allocator<void, local_elem_number, 18U>;
 
   using local_wide_decimal_type =
-    math::wide_decimal::decwide_t<wide_decimal_digits10, local_limb_type, local_allocator_type, double, std::int32_t>;
+    math::wide_decimal::decwide_t<wide_decimal_digits10, local_limb_type, local_allocator_type, float, std::int32_t>;
 
   const std::clock_t start = std::clock();
 
   const local_wide_decimal_type my_pi =
-    math::wide_decimal::pi<wide_decimal_digits10, local_limb_type, local_allocator_type, double, std::int32_t>(example002d_pi_limb8_digits10_callback);
+    math::wide_decimal::pi<wide_decimal_digits10, local_limb_type, local_allocator_type, float, std::int32_t>(example002d_pi_limb8_digits10_callback);
 
   const std::clock_t stop = std::clock();
